refactor: update README, adjust types, and clean up code
- Translated README content from French to English and added badges for license and PRs.
- Changed `duration` type in the Lyrics model from `int` to `float`.
- Removed unnecessary version declaration in docker-compose file.
- Standardized quotes in ESLint, Next.js, and other configuration files.
- Added Prettier as a dependency and configured format script in package.json.
- Cleaned up various components and hooks for consistency and readability.
- Updated card preview and export components for improved functionality and styling.

+++ backend/app/models/lyrics.py
@@ -20,7 +20,7 @@ class Lyrics(BaseModel):
     track_name: str
     artist_name: str
     album_name: str | None = None
-    duration: int | None = None
+    duration: float | None = None
     instrumental: bool = False
     lines: list[LyricLine]
     synced: bool = False
